ChainDB: pass new block to ledger validation
In `ChainDB.addBlock`, we receive a new block and use it to try switch to a
longer chain. This will require applying the block to the ledger, which
requires reading the block from disk and parsing it again. This is right in
the critical path of (bulk) chain sync.

Since we have the block in memory, we can avoid the redundant read and pass it
to the validation code directly.

Note that when switching to a fork or when we can extend the new chain with
more blocks after the new block, we'll still have to read blocks from disk in
order to validate them, but not the new block.